feat: add distributed tracing, prometheus and grafana (#4)
Adding otel sdk for distributed tracing and knitting with logging by
adding info to log config with trace id and if it was sampled. The
docker-compose just became huge, since all jaeger ecosystem was imported
and it actually needs some time to load cassandra and be brought up.

Locally, one can still use `make serve` and avoid the tracing ecosystem.

* chore(ci): do not upload to codecov on test