feat(api): add cache (#15)
* feat(api): add cache interface and valkey impl

The cache interface is composed of a class that forces
implementations to provide save, get, and delete methods.
Currently, the only implementation is valkey which has
a docker service limiting its memory. Application connects
to the cache instance via CACHE_HOST and CACHE_PORT
variables with CACHE_TTL_SECONDS set to 1h by default

* feat(api): cache in translate and detect

Now translate and detect routes will save the response
to cache and retrieve to reply faster to users. The cache
key is:
- translate:src_lang:tgt_lang:text
- detect:text

Where text is a sanitized version of the input text. By
sanitizations the app will do:
- Normalize Unicode characters (NFC form)
- Trim excess whitespaces
- Covert to lowercase